requirements.txt
requests
textblob
playsound>=1.2.2,<1.3.0
python-dotenv
schedule
colorama
pytz
tzlocal